# Optimize `signum` function (3-25x faster)