@findepi 2024-10-12
Are we going to remove all usages of `make_function_scalar_inputs_return_type`?