@simonvandel 2024-10-12
> Are we going to remove all usages of `make_function_scalar_inputs_return_type`?

If the replacement always has same or better performance, I don't see why not