@findepi 2024-10-12
Makes sense to me!
If this is an already established knowledge that replacing `make_function_scalar_inputs_return_type` with `unary` is an improvement, I don't see a need to benchmark this every time separately. IMO we could just do this change everywhere where the code pattern is same.