@simonvandel 2024-10-13
I don't mind adding the benchmarks, since they
- are easy to add
- will benefit the next contributor wanting to try some optimizations
- motivate me by quantifying performance of the changes I make