@alamb 2024-10-14
> Are we going to remove all usages of make_function_scalar_inputs_return_type?

Perhaps it could be a nice ticket to file / project ti organize